Move API webhook validation package
