+++ cmd/webhook/BUILD.bazel
@@ -14,7 +14,7 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/cmd/webhook",
     visibility = ["//visibility:private"],
     deps = [
-        "//pkg/apis/certmanager/validation/webhooks:go_default_library",
+        "//pkg/webhook/validation:go_default_library",
         "//vendor/github.com/openshift/generic-admission-server/pkg/cmd:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
     ],
